+++ internal/pkg/handler/authorization.go
@@ -2,8 +2,12 @@ package handler
 
 import (
 	"context"
+	"crypto/ecdsa"
+	"crypto/x509"
 
+	"github.com/dgrijalva/jwt-go/v4"
 	authz "github.com/koverto/authorization/api"
+	"github.com/koverto/authorization/pkg/claims"
 
 	"github.com/koverto/errors"
 	"github.com/koverto/micro"
@@ -13,11 +17,13 @@ import (
 type Authorization struct {
 	*Config
 	*micro.Service
-	client mongo.Client
+	client     mongo.Client
+	privateKey *ecdsa.PrivateKey
 }
 
 type Config struct {
-	MongoUrl string `json:"mongourl"`
+	MongoUrl   string `json:"mongourl"`
+	PrivateKey string `json:"privatekey"`
 }
 
 func New(conf *Config, service *micro.Service) (*Authorization, error) {
@@ -30,17 +36,25 @@ func New(conf *Config, service *micro.Service) (*Authorization, error) {
 		return nil, err
 	}
 
-	return &Authorization{conf, service, client}, nil
+	pkey, err := x509.ParseECPrivateKey([]byte(conf.PrivateKey))
+	if err != nil {
+		return nil, err
+	}
+
+	return &Authorization{conf, service, client, pkey}, nil
 }
 
-func (a *Authorization) Create(ctx context.Context, in *authz.TokenClaims, out *authz.Token) error {
+func (a *Authorization) Create(ctx context.Context, in *authz.TokenRequest, out *authz.Token) (err error) {
+	c := claims.New(in.GetUserID())
+	token := jwt.NewWithClaims(jwt.SigningMethodES256, c)
+	out.Token, err = token.SignedString(a.privateKey)
+	return err
+}
+
+func (a *Authorization) Validate(ctx context.Context, in *authz.Token, out *authz.TokenResponse) error {
 	return errors.NotImplemented(a.ID)
 }
 
-func (a *Authorization) Validate(ctx context.Context, in *authz.Token, out *authz.TokenClaims) error {
-	return errors.NotImplemented(a.ID)
-}
-
-func (a *Authorization) Invalidate(ctx context.Context, in *authz.TokenClaims, out *authz.TokenClaims) error {
+func (a *Authorization) Invalidate(ctx context.Context, in *authz.Token, out *authz.TokenResponse) error {
 	return errors.NotImplemented(a.ID)
 }