+++ CHANGELOG.md
@@ -7,5 +7,10 @@ The format is based on [Keep a Changelog][], and this project adheres to
 
 ## Unreleased
 
+### Added
+
+- Protobuf API
+- `Authorization.Create`
+
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html